fix autocomplete, fix style of login button and page
