+++ app/db.ts
@@ -1,3 +1,5 @@
 import PocketBase from "pocketbase";
 
-export const db = new PocketBase(process.env.POCKETBASE_URL);
+const { POCKETBASE_URL } = process.env;
+console.log(POCKETBASE_URL)
+export const db = new PocketBase("http://127.0.0.1:8090");
